+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## 4.1.2 (unreleased)
+## 4.2.0 (2022-08-05)
 
 ### Feature